Move UIImage test helpers to a separated file
