+++ hurb-test/hurb-testTests/Helper/UIImage+TestHelpers.swift
@@ -1,0 +1,22 @@
+//
+//  UIImage+TestHelpers.swift
+//  hurb-testTests
+//
+//  Created by Tulio Parreiras on 20/09/20.
+//  Copyright © 2020 Tulio Parreiras. All rights reserved.
+//
+
+import UIKit
+
+extension UIImage {
+    static func make(withColor color: UIColor) -> UIImage {
+        let rect = CGRect(x: 0, y: 0, width: 1, height: 1)
+        UIGraphicsBeginImageContext(rect.size)
+        let context = UIGraphicsGetCurrentContext()!
+        context.setFillColor(color.cgColor)
+        context.fill(rect)
+        let img = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+        return img!
+    }
+}
